requirements.txt
MTCNN
numpy
pandas
Pillow
tensorboardX
torch
torchvision
tqdm
tensorflow
opencv-python
